디버그 모드로 컬라이더, a*영역, 오토 타겟팅 영역 등등 껐다 키기,  캐릭터 죽일 경우 이펙트 UI 나타나기

+++ WinAPI/CGTA_RunawayState.cpp
@@ -107,8 +107,8 @@ void CGTA_RunawayState::Update()
 		wstring strHelpSoundPath = Sound_Help + std::to_wstring((rand() % Sound_Help_Len) + 1);
 		CSound* pSound = CResourceManager::GetInstance()->GetSound(strHelpSoundPath, strHelpSoundPath);
 		pSound->Play();
-		m_fHelpSoundMaxCoolTime = rand() % 10 + 3;
-		m_fHelpSoundCoolTime = 0;
+		m_fHelpSoundMaxCoolTime = (float)(rand() % 10 + 3);
+		m_fHelpSoundCoolTime = 0.f;
 	}
 }
 
@@ -123,7 +123,7 @@ void CGTA_RunawayState::Start()
 	}
 
 	m_fHelpSoundCoolTime = 0;
-	m_fHelpSoundMaxCoolTime = rand() % 10 + 3;
+	m_fHelpSoundMaxCoolTime = float(rand() % 10 + 3);
 }
 
 void CGTA_RunawayState::End()
